Refactor CommandArguments
