+++ src/CodeGenerator/Domain/CommandArgumentsParser.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Gacela\CodeGenerator\Domain;
 
 use Gacela\CodeGenerator\Domain\Exception\CommandArgumentsException;
-use Gacela\CodeGenerator\Domain\ValueObject\CommandArguments;
 use InvalidArgumentException;
 
 final class CommandArgumentsParser
